Add `auth_anonymous` option
This is just exposing skopeo's `--no-creds`.  I would like
to flip the ostree-as-system case to use this by default, because
I don't think the host OS upgrades should read from `/root` by
default.

Instead, I plan to change the ostree stack to support something like
`/etc/ostree/container-auth.json` so it can be distinct from
what the "root user"'s config is.

Further, the config needs to be in `/etc` because `/root` is on
`/var` on ostree systems, and we want to support transactional
updates to the pull config.

(Anyways, that's all mostly ostree stuff, this is just exposing
 the option)