Use compile time rather than runtime detection for v0_2_3
I thought my initial code to dynamically detect the proxy
version was cool - it maximizes flexibility.  I think it is a pattern
we will want in the future.

However, what I realized when doing some work related to this in
ostree-ext is that for this particular API, we need to make it a hard
dependency.

So we might as well do the much simpler thing and make it a compile
time feature.  Then all we need to do is change the runtime code
to match on the build-time selected semver.