Test that CI successfully fails
I think a good pattern is to keep an open (draft) pull request
that *should* fail CI, and verify that it actually does.

It's surprisingly easily to e.g. have CI not actually test the
code.  In this case we have our "reverse dependency testing"
which patches the crate to use the local code, and that could
easily not work.