Change test case to match error via types, not strings
Minor followup to the previous change; part of the idea of having
the error types here is so we can *also* dig through the `SkopeoSpawnError`
and precisely match against the `std::io::ErrorKind::NotFound`,
and not rely on its stringified form. This is what we'd expect
other callers to do.

This also avoids a double-match.

Signed-off-by: Colin Walters <walters@verbum.org>