imageproxy: make ImageProxy: Send + Sync
The internal detail of how we handle the exit of the child process leads
to the entire ImageProxy struct not being Send, which means that it
can't be used in cross-thread Futures with tokio.  Fortunately it's a
very simple fix.

Add some trivial static testing to make sure this doesn't regress.  Add
it for OpenedImage as well: it's currently trivially Send+Sync on
account of being a wrapper around an integer, but this might change in
the future.

This change has been cross-checked against both bootc (ostree-ext) and
composefs-rs and doesn't cause any issues.

Signed-off-by: Allison Karlitskaya <allison.karlitskaya@redhat.com>